Oce updates

+++ web/src/main/java/org/devgateway/ocds/web/rest/controller/request/GenericPagingRequest.java
@@ -3,11 +3,10 @@
  */
 package org.devgateway.ocds.web.rest.controller.request;
 
+import io.swagger.annotations.ApiModelProperty;
 import org.hibernate.validator.constraints.Range;
 
 import javax.validation.constraints.Min;
-
-import io.swagger.annotations.ApiModelProperty;
 
 /**
  * @author mpostelnicu
@@ -58,7 +57,7 @@ public class GenericPagingRequest {
         this.pageSize = size;
     }
 
-    public Integer getSkip() {
-        return pageNumber * pageSize;
+    public Long getSkip() {
+        return (long) (pageNumber * pageSize);
     }
 }
